requirements.txt
requests
bs4
Pyyaml
numpy
matplotlib
syllables